Fixed shadowwarnings

+++ Aufgaben/Aufgaben/FiveDegreeInterpolation.cpp
@@ -4,7 +4,7 @@
 float CFiveDegreeInterpolation::Interpolation(float functionvalue)
 {
 	// S(t) = 10*t^3 - 15*t^4 + 6*t^5
-	return 10 * std::pow(functionvalue, 3)
-		- 15 * std::pow(functionvalue, 4)
-		+ 6 * std::pow(functionvalue, 5);
+	return 10 * std::powf(functionvalue, 3.0f)
+		- 15 * std::powf(functionvalue, 4.0f)
+		+ 6 * std::powf(functionvalue, 5.0f);
 }
