+++ Aufgaben/Aufgaben/ThreeDegreeInterpolation.cpp
@@ -3,6 +3,6 @@
 
 float CThreeDegreeInterpolation::Interpolation(float functionvalue)
 {
-    return 3 * std::pow(functionvalue, 2)
-        - 2 * std::pow(functionvalue, 3);
+    return 3 * std::powf(functionvalue, 2.0f)
+        - 2 * std::powf(functionvalue, 3.0f);
 }
